+++ R/get_var_corr_.R
@@ -61,8 +61,10 @@ if("factor" %in% sapply(df,class)){
   if(subset_df==TRUE){
     res[res$Comparison_Var %in%subset_cols[[1]] &
           res$Other_Var %in% subset_cols[[2]],]
+   
   }
   else{
+ 
     res
   }
 } else{
@@ -80,6 +82,7 @@ if("factor" %in% sapply(df,class)){
   if(subset_df==TRUE){
     res[res$Comparison_Var %in%subset_cols[[1]] &
           res$Other_Var %in% subset_cols[[2]],]
+   
   }
   else{
     res